+++ src/components/Account/index.ts
@@ -1,3 +1,2 @@
-export { default as Container } from './Container';
 export { default as ChangePassWord } from './ChangePassWord';
 export { default as CreateWiki } from './CreateWiki';
