--- src/components/Auth/index.ts
@@ -1,3 +1,0 @@
-export { default as AuthContainer } from './AuthContainer';
-export { default as AuthSwitchPrompt } from './AuthSwitchPrompt';
-export { default as AuthInputWithLabel } from './AuthInputWithLabel';
