+++ src/components/common/Form/InputWithLabel.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { FieldValues } from 'react-hook-form';
 import { Label, Input } from '@/components/common/Form';
 import { InputWithLabelProps } from '@/types/authFormType';
-import UnLockIcon from '@/components/Auth/UnLock';
-import LockIcon from '@/components/Auth/Lock';
+import UnLockIcon from '@/components/common/Form/UnLock';
+import LockIcon from '@/components/common/Form/Lock';
 import useBoolean from '@/hooks/useBoolean';
 
 const AuthInputWithLabel = <T extends FieldValues>({
@@ -17,11 +17,11 @@ const AuthInputWithLabel = <T extends FieldValues>({
   ...props
 }: InputWithLabelProps<T>) => {
   const [inputType, setInputType] = useState(type);
-  const { value: showPassWord, handleToggle: togglePassword } = useBoolean();
+  const { value: password, handleToggle: togglePassword } = useBoolean();
 
   const handleIconClick = () => {
     togglePassword();
-    setInputType(showPassWord ? 'password' : 'text');
+    setInputType(password ? 'password' : 'text');
   };
 
   const hasError = !!errors[name];
@@ -29,7 +29,7 @@ const AuthInputWithLabel = <T extends FieldValues>({
 
   return (
     <div className="mb-6 flex flex-col gap-2.5">
-      <div className="authContainer">
+      <div className="formContainer">
         <Input
           id={id}
           type={inputType}
@@ -40,7 +40,7 @@ const AuthInputWithLabel = <T extends FieldValues>({
         <Label htmlFor={id} label={label} className={`label ${hasError && 'labelError'}`} />
         {type === 'password' && (
           <span className={`checkPassword ${hasError ? 'top-1/3' : ''}`} onClick={handleIconClick}>
-            {showPassWord ? <UnLockIcon /> : <LockIcon />}
+            {password ? <UnLockIcon /> : <LockIcon />}
           </span>
         )}
         {hasError && <span className="errorMessage">{errorMessages}</span>}
