+++ src/components/common/Form/index.ts
@@ -1,2 +1,7 @@
 export { default as Input } from './Input';
 export { default as Label } from './Label';
+export { default as Container } from './Container';
+export { default as InputWithLabel } from './InputWithLabel';
+export { default as SwitchPrompt } from './SwitchPrompt';
+export { default as Lock } from './Lock';
+export { default as UnLock } from './UnLock';
