+++ src/pages/mypage.tsx
@@ -1,11 +1,11 @@
-import { Container, ChangePassWord, CreateWiki } from '@/components/Account';
+import { ChangePassWord, CreateWiki } from '@/components/Account';
 
 const MyPage = () => {
   return (
-    <Container title="계정 설정">
+    <>
       <ChangePassWord />
       <CreateWiki />
-    </Container>
+    </>
   );
 };
